# api.yaml
openapi: 3.0.0
info:
  title: OMS Backend (MVP)
  version: 1.0.0
  description: Minimal backend API for sending trading orders to Centroid through an internal FIX engine.

servers:
  - url: http://localhost:8000

paths:

  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: OK

  /orders:
    post:
      summary: Create a new trading order
      description: >
        Creates an order in the local OMS. The order is persisted and then
        routed to the FIX engine which will forward it to Centroid.
      parameters:
        - in: header
          name: X-Client-Id
          required: false
          description: Optional. Client ID can be sent in body as clientId or in this header.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreateRequest'
      responses:
        '201':
          description: Order accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '422':
          description: Validation error

    get:
      summary: List orders
      description: Requires a client identifier provided either as query ?clientId= or header X-Client-Id. Returns 400 if neither provided.
      parameters:
        - in: query
          name: clientId
          schema:
            type: string
        - in: header
          name: X-Client-Id
          required: false
          schema:
            type: string
        - in: query
          name: symbol
          schema:
            type: string
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '400':
          description: Missing clientId (query or X-Client-Id header)

  /orders/{orderId}:
    get:
      summary: Get a single order by ID
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Not found

  /orders/{orderId}/cancel:
    post:
      summary: Cancel an order
      description: >
        Requests cancellation. The FIX engine will send FIX Cancel (35=F)
        to Centroid and update status when the ExecutionReport arrives.
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Cancel request accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found

  /positions:
    get:
      summary: Get current positions for a client
      description: Client can be provided via query ?clientId= or header X-Client-Id. Defaults to demo-client-1 if missing.
      parameters:
        - in: query
          name: clientId
          schema:
            type: string
        - in: header
          name: X-Client-Id
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Positions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Position'

  /v1/dashboard:
    get:
      summary: Dashboard
      description: Returns withdrawable balance and strategies for the client.
      parameters:
        - in: query
          name: clientId
          schema:
            type: string
        - in: header
          name: X-Client-Id
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Dashboard
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dashboard'

  /v1/deposits:
    get:
      summary: List deposits
      parameters:
        - in: query
          name: clientId
          schema:
            type: string
        - in: header
          name: X-Client-Id
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Deposits list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Deposit'
    post:
      summary: Create deposit
      parameters:
        - in: query
          name: clientId
          schema:
            type: string
        - in: header
          name: X-Client-Id
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepositCreate'
      responses:
        '200':
          description: Deposit created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deposit'

  /v1/withdrawals:
    get:
      summary: List withdrawals
      parameters:
        - in: query
          name: clientId
          schema:
            type: string
        - in: header
          name: X-Client-Id
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Withdrawals list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Withdrawal'
    post:
      summary: Create withdrawal
      parameters:
        - in: query
          name: clientId
          schema:
            type: string
        - in: header
          name: X-Client-Id
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WithdrawalCreate'
      responses:
        '200':
          description: Withdrawal created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Withdrawal'

components:

  schemas:

    OrderCreateRequest:
      type: object
      required:
        - symbol
        - side
        - type
        - qty
      properties:
        clientId:
          type: string
        symbol:
          type: string
          example: XAUUSD
        side:
          type: string
          enum: [BUY, SELL]
        type:
          type: string
          enum: [MARKET, LIMIT]
        qty:
          type: number
        price:
          type: number
          nullable: true
        timeInForce:
          type: string
          enum: [GTC, DAY, IOC, FOK]
          default: GTC

    Order:
      type: object
      properties:
        id:
          type: string
        clientId:
          type: string
        symbol:
          type: string
        side:
          type: string
        type:
          type: string
        qty:
          type: number
        price:
          type: number
          nullable: true
        status:
          type: string
          enum:
            - NEW
            - PENDING_SEND
            - SENT
            - PARTIALLY_FILLED
            - FILLED
            - REJECTED
            - CANCELED
        cumQty:
          type: number
        filledQty:
          type: number
        avgPx:
          type: number
          nullable: true
        rejectReason:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string

    Position:
      type: object
      properties:
        clientId:
          type: string
        symbol:
          type: string
        netQty:
          type: number
        avgPx:
          type: number
        unrealizedPnl:
          type: number

    Dashboard:
      type: object
      properties:
        saldo_retirable:
          type: number
        estrategias:
          type: array
          items:
            type: object

    DepositCreate:
      type: object
      required:
        - amount
      properties:
        amount:
          type: number
        currency:
          type: string
          default: USD
        payment_method:
          type: string
          nullable: true
        provider:
          type: string
          nullable: true
        status:
          type: string
          default: completed

    Deposit:
      type: object
      properties:
        id:
          type: string
        amount:
          type: number
        currency:
          type: string
        payment_method:
          type: string
          nullable: true
        provider:
          type: string
          nullable: true
        status:
          type: string
        created_at:
          type: string

    WithdrawalCreate:
      type: object
      required:
        - amount
      properties:
        amount:
          type: number
        currency:
          type: string
          default: USD
        status:
          type: string
          default: completed

    Withdrawal:
      type: object
      properties:
        id:
          type: string
        amount:
          type: number
        currency:
          type: string
        status:
          type: string
        created_at:
          type: string